refactor: Tabs 컴포넌트가 Tab 컴포넌트만 props로 받도록 수정
